+++ .docker/Dockerfile
@@ -11,15 +11,14 @@ ENV TZ=Europe/Paris
 # - Persistent / Runtime dependencies
 RUN set -eux; \
     apk add --no-cache \
-    curl \
-    git \
-    nginx \
-    supervisor \
-    tzdata \
-    unzip \
-    dcron \
-    logrotate \
-    jq;
+        curl \
+        nginx \
+        supervisor \
+        tzdata \
+        unzip \
+        dcron \
+        logrotate \
+        jq;
 
 # - User / Group "loxya"
 RUN set -eux; \
@@ -40,34 +39,30 @@ COPY --from=mlocati/php-extension-installer:2.8.5 /usr/bin/install-php-extension
 RUN \
     --mount=type=cache,target=/var/cache/apt \
     install-php-extensions \
-    apcu \
-    bcmath \
-    curl \
-    dom \
-    fileinfo \
-    ftp \
-    gd \
-    gettext \
-    iconv \
-    imagick \
-    intl \
-    json \
-    mbstring \
-    mysqli \
-    opcache \
-    openssl \
-    pcre \
-    pdo \
-    pdo_mysql \
-    soap \
-    sockets \
-    ssh2 \
-    xml \
-    xsl \
-    zip
-
-# - Git Configuration
-COPY ./config/gitconfig /etc/gitconfig
+        apcu \
+        bcmath \
+        curl \
+        dom \
+        fileinfo \
+        gd \
+        gettext \
+        iconv \
+        imagick \
+        intl \
+        json \
+        mbstring \
+        mysqli \
+        opcache \
+        openssl \
+        pcre \
+        pdo \
+        pdo_mysql \
+        soap \
+        sockets \
+        ssh2 \
+        xml \
+        xsl \
+        zip
 
 # - PHP Configuration
 COPY ./config/php.ini ${PHP_INI_DIR}/conf.d/custom.ini
@@ -75,22 +70,22 @@ COPY ./config/php-fpm.conf /usr/local/etc/php-fpm.conf
 RUN set -eux; \
     mv ${PHP_INI_DIR}/php.ini-production ${PHP_INI_DIR}/php.ini; \
     rm -r \
-    ${PHP_INI_DIR}/php.ini-development \
-    /usr/local/etc/php-fpm.conf.default \
-    /usr/local/etc/php-fpm.d;
+        ${PHP_INI_DIR}/php.ini-development \
+        /usr/local/etc/php-fpm.conf.default \
+        /usr/local/etc/php-fpm.d;
 
 # - Nginx Configuration
 COPY ./config/nginx.conf /etc/nginx/nginx.conf
 COPY ./config/vhosts/default.conf /etc/nginx/vhosts/default.conf
 RUN set -eux; \
     rm -rf \
-    /var/www \
-    /etc/nginx/modules \
-    /etc/nginx/http.d; \
+        /var/www \
+        /etc/nginx/modules \
+        /etc/nginx/http.d; \
     chown -R loxya:loxya \
-    /etc/nginx/vhosts \
-    /var/lib/nginx \
-    /var/log/nginx;
+        /etc/nginx/vhosts \
+        /var/lib/nginx \
+        /var/log/nginx;
 
 # - Supervisor configuration
 COPY ./config/supervisord.conf /etc/supervisord.conf
@@ -98,14 +93,14 @@ COPY ./config/supervisord.conf /etc/supervisord.conf
 # - Loxya base directories
 RUN set -eux; \
     mkdir -p \
-    /opt/loxya \
-    /var/loxya/logs; \
+        /opt/loxya \
+        /var/loxya/logs; \
     chown -R loxya:loxya \
-    /opt/loxya \
-    /var/loxya; \
+        /opt/loxya \
+        /var/loxya; \
     chmod -R 2770 \
-    /opt/loxya \
-    /var/loxya;
+        /opt/loxya \
+        /var/loxya;
 
 EXPOSE 80
 
@@ -130,32 +125,32 @@ FROM app-base AS app-release
 
 RUN set -eux; \
     mkdir -p \
-    /etc/loxya \
-    /tmp/loxya \
-    /var/loxya/data \
-    /var/loxya/cache \
-    /var/loxya/data/bills \
-    /var/loxya/data/imports \
-    /var/loxya/data/materials; \
+        /etc/loxya \
+        /tmp/loxya \
+        /var/loxya/data \
+        /var/loxya/cache \
+        /var/loxya/data/bills \
+        /var/loxya/data/imports \
+        /var/loxya/data/materials; \
     chown -R loxya:loxya \
-    /etc/loxya \
-    /tmp/loxya \
-    /var/loxya; \
+        /etc/loxya \
+        /tmp/loxya \
+        /var/loxya; \
     chmod -R 2770 \
-    /etc/loxya \
-    /tmp/loxya \
-    /var/loxya;
+        /etc/loxya \
+        /tmp/loxya \
+        /var/loxya;
 
 # - Copie des sources
 COPY --chown=loxya:loxya --from=source . /opt/loxya
 RUN set -eux; \
     rm -rf \
-    /opt/loxya/data \
-    /opt/loxya/src/var/logs \
-    /opt/loxya/src/var/cache \
-    /opt/loxya/src/var/tmp \
-    /opt/loxya/src/App/Config/install.json \
-    /opt/loxya/src/App/Config/settings.json; \
+        /opt/loxya/data \
+        /opt/loxya/src/var/logs \
+        /opt/loxya/src/var/cache \
+        /opt/loxya/src/var/tmp \
+        /opt/loxya/src/App/Config/install.json \
+        /opt/loxya/src/App/Config/settings.json; \
     ln -snf /var/loxya/data /opt/loxya/data; \
     ln -snf /tmp/loxya /opt/loxya/src/var/tmp; \
     ln -snf /var/loxya/logs /opt/loxya/src/var/logs; \
