--- .docker/config/gitconfig
@@ -1,4 +1,0 @@
-[safe]
-  directory = /opt/loxya
-  directory = /opt/loxya/server/vendors/bigfish/pdf417
-  directory = /opt/loxya/server/vendors/respect/validation
