--- client/.node-version
@@ -1,1 +1,0 @@
-v24.8.0
