--- client/src/themes/default/pages/Settings/Global/pages/Companies/index.scss
@@ -1,1 +1,0 @@
-@use '~@/themes/default/style/globals';
