--- client/src/themes/default/pages/Settings/Global/pages/Companies/modals/index.scss
@@ -1,73 +1,0 @@
-@use '~@/themes/default/style/globals';
-
-.ModalBillingCompanyEdition {
-  $block: &;
-
-  .FormField .FormField__label {
-    flex: 0 0 calc(var(--label-width) + 10px);
-  }
-
-  &__header {
-    display: flex;
-    align-items: center;
-    padding: 0.7rem 1rem;
-    background-color: globals.$bg-color-emphasis;
-    gap: globals.$spacing-medium;
-
-    &__title {
-      flex: 1;
-      overflow: hidden;
-      margin: 0;
-      font-size: 1.3rem;
-      white-space: nowrap;
-      text-overflow: ellipsis;
-    }
-
-    &__close-button {
-      flex: 0 0 auto;
-    }
-  }
-
-  &__body {
-    padding: globals.$spacing-large globals.$spacing-medium;
-  }
-
-  &__footer {
-    display: flex;
-    justify-content: center;
-    padding: 1.5rem globals.$spacing-medium;
-    background-color: rgb(0 0 0 / 10%);
-    gap: globals.$spacing-small;
-  }
-
-  &__locality {
-    display: flex;
-    margin: globals.$spacing-medium 0;
-
-    #{$block}__city {
-      flex: 1;
-      margin: 0 0 0 10px;
-      .FormField__label {
-        flex: 0 auto;
-      }
-    }
-  }
-  &__legal-numbers {
-    .Fieldset__body  {
-      display: grid;
-      grid-template-columns: repeat(2, 1fr);
-
-      .FormField__label {
-        flex: 0 auto;
-      }
-    }
-  }
-  &__logo {
-    width: var(--input-width);
-
-    .InputImage {
-      margin: auto;
-    }
-  }
-}
-
